initialise workflow
