+++ plugins/scaffolder-frontend-workflow/src/hooks/index.ts
@@ -1,2 +1,1 @@
 export * from './useStepper';
-export * from './useRunWorkflow';
